Issue #1095: Assistants compat follow-ups
