+++ tests/live/README.md
@@ -63,6 +63,25 @@ Notes:
 | `npm run test:live:openai` | Run live tests with OpenAI |
 | `npm run test:live:google` | Run live tests with Google |
 
+## LLM Provider Credentials
+
+The live runner loads `./.env` and hard-fails when required env vars are missing (see `tests/live/required-env.ts`).
+
+Common env vars:
+
+```bash
+OPENROUTER_API_KEY=...
+ANTHROPIC_API_KEY=...
+OPENAI_API_KEY=...
+OPENAI_ASSISTANTS_ASSISTANT_ID=asst_...
+GEMINI_API_KEY=...
+
+# Azure OpenAI Assistants (when selected)
+AZURE_OPENAI_API_KEY=...
+AZURE_OPENAI_ENDPOINT=...
+OPENAI_API_VERSION=...
+```
+
 ## Realtime live tests
 
 Realtime live tests validate the provider-agnostic realtime session contract (audio/text in, audio/transcripts out, tools, barge-in). These tests only run when:
